pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.45.2",
    "autogen>=0.7.4",
    "autogen-agentchat[anthropic]>=0.4.7",
    "autogen-ext[anthropic,openai,semantic-kernel-anthropic,web-surfer]>=0.4.7",
    "autogenstudio>=0.4.1.5",
    "dotenv>=0.9.9",
    "gitpython>=3.1.44",
    "httpx>=0.28.1",
    "langchain>=0.3.19",
    "langchain-anthropic>=0.3.7",
    "langchain-core>=0.3.35",
    "langgraph>=0.2.73",
    "langsmith>=0.3.8",
    "logfire>=3.6.2",
    "mcp[cli]>=1.2.1",
    "pydantic>=2.10.6",
    "pydantic-ai[logfire]>=0.0.24",
    "smolagents>=1.9.2",
    "typing-extensions>=4.12.2",
]
